refactor(filterbar-select): rename FilterBarSelect files
