+++ packages/components/src/FilterBar/subcomponents/FilterBarSelect/index.ts
@@ -1,0 +1,1 @@
+export * from "./FilterBarSelect"
