--- packages/components/src/FilterBar/subcomponents/FilterSelect/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterSelect"
